api/api_drf/requirements.txt
django==5.0
djangorestframework==3.14.0
django-filter==23.5
psycopg2-binary==2.9.9
matplotlib==3.8.2
momepy==0.7.0
neo4j==5.15.0
neomodel==5.2.1
networkx==3.2.1
sqids==0.4.1
shapely==2.0.3
wntr==1.1.0
